refactor Pids() function in Linux Parser class to Processes() class in Processor, so we always get a live reading of the active processes on the system.
reorganize headers accordingly to make the above work. If it isn't clear by now the goal is to ultimately remove the Linux Parser class and have each System, Processor, and Process class handle and retrieve the relevant information in its area, housing the logic required to do so so that we can adhere to encapsulation. Linux Parser is redundant and clutters the project, no reason for it to stick around and just increase call overhead.

removeDead() is a helper function to that erases all elements from the processes vector that are not found by the Processes() function call. So in other words, a killed process that we have not removed from our vector. Using this along with the find_if call in Processes(), we can update and initialize new unrecorded Processes only, while removing dead ones, instead of creating a new vec everytime.