implement FindUid() and two overloaded definitions of a User() function in the Process class to retrieve the process uid from the proc/{pid}/status file, and then set/retrieve the userName property respectively
create a Processor constructor to read the contents of the large /etc/passwd file into an std::unordered_map for easy process Uid lookup. Loading the file continuously, using regex, or even using lots of popen calls was causing the program to hang and suffer from memory leaks, so this turned out to be the most performant way!